requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==1.10.14
PyYAML==6.0.1
pillow==10.2.0
pdfplumber==0.10.3
python-docx==0.8.11
pytesseract==0.3.10
pytest==8.1.1
